Commit of Cassandra plugin
This plugin is functional, but the main reason why I am committing this is because of the jar file.
The jar file has been re-compiled to allow passing of object args using the full qualifier for the
object. This was done initially because the munin core jar file could not perform this action, so
certain jmx objects were unqueryable, with this jar file all jmx objects can be fetched, assuming
you know the qualifier to grab.